CloudGRC-AI-Portfolio/CloudGRC-AI/requirements.txt
boto3>=1.34.0
google-cloud-storage>=2.14.0
google-cloud-iam>=2.14.0
google-cloud-logging>=3.9.0
azure-mgmt-storage>=21.1.0
azure-mgmt-network>=25.1.0
azure-mgmt-monitor>=6.0.0
azure-identity>=1.16.0
azure-mgmt-authorization>=4.0.0
openai>=1.30.0
streamlit>=1.35.0
reportlab>=4.2.0
pandas>=2.2.0
plotly>=5.22.0
python-dotenv>=1.0.1
rich>=13.7.0
requests>=2.32.0
